provide a base implementation for AnnotationDeducer based on annotations
