+++ src/main/kotlin/com/tegonal/minimalist/providers/impl/DefaultAnnotationDataDeducer.kt
@@ -1,14 +1,9 @@
 package com.tegonal.minimalist.providers.impl
 
-import com.tegonal.minimalist.export.org.junit.platform.commons.util.AnnotationUtils
 import com.tegonal.minimalist.providers.AnnotationData
-import com.tegonal.minimalist.providers.AnnotationDataDeducer
 import com.tegonal.minimalist.providers.ArgsSourceOptions
+import com.tegonal.minimalist.providers.BaseAnnotationDataDeducer
 import com.tegonal.minimalist.providers.fromOptions
-import com.tegonal.minimalist.providers.merge
-import java.lang.reflect.AnnotatedElement
-import java.lang.reflect.Method
-import kotlin.jvm.optionals.getOrNull
 
 /**
  * !! No backward compatibility guarantees !!
@@ -16,23 +11,10 @@ import kotlin.jvm.optionals.getOrNull
  *
  * @since 2.0.0
  */
-class DefaultAnnotationDataDeducer : AnnotationDataDeducer {
+class DefaultAnnotationDataDeducer : BaseAnnotationDataDeducer<ArgsSourceOptions>(ArgsSourceOptions::class) {
 
-	override fun deduce(
-		testMethod: Method,
-		argsSourceMethodName: String
-	): AnnotationData? {
-		val classData = deduceData(testMethod.declaringClass, argsSourceMethodName)
-		val methodData = deduceData(testMethod, argsSourceMethodName)
-		return when {
-			classData != null && methodData != null -> classData.merge(methodData)
-			classData != null && methodData == null -> classData
-			methodData != null -> methodData
-			else -> null
-		}
-	}
-
-	private fun deduceData(element: AnnotatedElement, argsSourceMethodName: String): AnnotationData? =
-		AnnotationUtils.findAnnotation(element, ArgsSourceOptions::class.java)
-			.map { AnnotationData.fromOptions(argsSourceMethodName, it) }.getOrNull()
+	override fun annotationToAnnotationData(
+		argsSourceMethodName: String,
+		annotation: ArgsSourceOptions
+	): AnnotationData = AnnotationData.fromOptions(argsSourceMethodName, annotation)
 }
